backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.9.1
aiohttp==3.11.18
aiofiles==24.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.20
python-dotenv==1.0.0
yt-dlp==2025.3.31
